Move get user tricks function to a handler file
